Add metastore_facets, update test
