--- modules/dkan_common/tests/src/Functional/DkanCommonUpdatePathTest.php
@@ -1,59 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Drupal\Tests\dkan_common\Functional;
-
-use Drupal\FunctionalTests\Update\UpdatePathTestBase;
-
-/**
- * Tests update functions for the metastore module.
- *
- * @group metastore
- * @group update
- * @group functional3
- */
-class DkanCommonUpdatePathTest extends UpdatePathTestBase {
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function setDatabaseDumpFiles() {
-    $this->databaseDumpFiles = [
-      dirname(__DIR__, 5) . '/tests/fixtures/update/update-2.20.0.php.gz',
-    ];
-  }
-
-  /**
-   * Test that common settings are copied to dkan_common on install.
-   */
-  public function testRenameUpdate(): void {
-    // Drop cache container file to avoid issues with module moving.
-    \Drupal::database()->truncate('cache_container')->execute();
-
-    $this->drupalGet('<front>');
-    $this->assertSession()->pageTextContains('Log in');
-
-    $config = \Drupal::configFactory()->getEditable('common.settings');
-    $this->assertFalse($config->get('always_use_existing_local_perspective'));
-    // Assert common module is installed
-    $this->assertTrue(\Drupal::moduleHandler()->moduleExists('common'));
-
-    // Set always_use_existing_local_perspective to TRUE to ensure it survives
-    // the update.
-    $config->set('always_use_existing_local_perspective', TRUE);
-    $config->save();
-
-    // Get a baseline for pre-install.
-    $this->assertTrue($config->get('always_use_existing_local_perspective'));
-
-    \Drupal::service('module_installer')->install(['dkan_common']);
-    $this->runUpdates();
-
-    // Assert setting copied over
-    $newConfig = \Drupal::configFactory()->getEditable('dkan_common.settings');
-    $this->assertTrue($newConfig->get('always_use_existing_local_perspective'));
-
-  }
-
-}
