+++ modules/dkan_metastore/modules/dkan_metastore_facets/dkan_metastore_facets.info.yml
@@ -1,0 +1,7 @@
+name: DKAN Metastore Facets (transitional)
+description: Provides facet blocks for DKAN Metastore Search.
+type: module
+core_version_requirement: ^10.4 || ^11
+dependencies:
+  - dkan:metastore_facets
+package: DKAN 4.x Transition
